refactor(mobile): remove no longer used layout code and redundant NetworkBadge usages
