--- apps/mobile/src/components/headers/components/header-leather-logo.tsx
@@ -1,9 +1,0 @@
-import { Box, LeatherLogomarkIcon } from '@leather.io/ui/native';
-
-export function HeaderLeatherLogo() {
-  return (
-    <Box p="3">
-      <LeatherLogomarkIcon />
-    </Box>
-  );
-}
