+++ apps/mobile/src/components/sheets/full-height-sheet/full-height-sheet-header.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
-import { HeaderTitleWithSubtitle } from '../../headers/components/header-title-with-subtitle';
-import { HeaderLayout } from '../../headers/header.layout';
+import { HeaderTitleWithSubtitle } from '@/components/screen/screen-header/components/header-title-with-subtitle';
+import { HeaderLayout } from '@/components/screen/screen-header/components/header.layout';
 
 interface FullHeightSheetHeaderProps {
   title: string;
