+++ apps/mobile/src/features/receive/screens/select-asset.tsx
@@ -1,8 +1,7 @@
-import { HeaderBackButton } from '@/components/headers/components/header-back-button';
+import { HeaderBackButton } from '@/components/screen/screen-header/components/header-back-button';
 import { FullHeightSheetHeader } from '@/components/sheets/full-height-sheet/full-height-sheet-header';
 import { FullHeightSheetLayout } from '@/components/sheets/full-height-sheet/full-height-sheet.layout';
 import { useReceiveFlowContext } from '@/features/receive/receive-flow-provider';
-import { NetworkBadge } from '@/features/settings/network-badge';
 import { TestId } from '@/shared/test-id';
 import { useBitcoinPayerAddressFromAccountIndex } from '@/store/keychains/bitcoin/bitcoin-keychains.read';
 import { useStacksSignerAddressFromAccountIndex } from '@/store/keychains/stacks/stacks-keychains.read';
@@ -76,7 +75,6 @@ export function SelectAsset() {
             leftElement={
               canGoBack ? <HeaderBackButton onPress={goBack} testID={TestId.backButton} /> : null
             }
-            rightElement={<NetworkBadge />}
           />
         }
       >
